+++ database/migrations/2022_03_20_122139_create_customer_billing_payment_logs_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('customer_billing_payment_logs', function (Blueprint $table) {
             $table->id();
             $table->integer('customer_id');
-            $table->integer('dry_weight');
+            $table->float('dry_weight');
             $table->float('payment_paid');
             $table->timestamps();
         });
